{"schema":"decodex/commit/1","summary":"split daemon and home interaction modules","authority":"manual"}
