+++ site/src/scripts/homeInteractions/bind.js
@@ -1,0 +1,9 @@
+function bindHomeInteractions() {
+  bindReveals();
+  bindPressStates();
+  bindHeroPointer();
+}
+
+document.addEventListener("astro:page-load", bindHomeInteractions);
+document.addEventListener("DOMContentLoaded", bindHomeInteractions);
+bindHomeInteractions();
