+++ site/src/scripts/homeInteractions/query.js
@@ -1,0 +1,3 @@
+function all(selector) {
+  return Array.prototype.slice.call(document.querySelectorAll(selector));
+}
