+++ site/src/scripts/homeInteractions/state.js
@@ -1,0 +1,1 @@
+const reduceMotion = window.matchMedia("(prefers-reduced-motion: reduce)");
